requirements.txt
torch>=1.9.0
transformers>=4.20.0
scikit-learn>=1.0.0
pandas>=1.3.0
numpy>=1.21.0
matplotlib>=3.4.0
seaborn>=0.11.0
tqdm>=4.62.0
PyPDF2>=3.0.0
pdfplumber>=0.7.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
# Output parsers and structured formatting
langchain>=0.1.0
langchain-core>=0.1.0
pydantic>=2.0.0
